Implement useLayoutEffect and other improvements
