+++ src/reducers/ComparisonSlice.ts
@@ -3,6 +3,7 @@ import { createSelector, createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { RootState } from '../common/store';
 import { Strings } from '../resources/Strings';
 import { CompareResultsItem, RevisionsHeader } from '../types/state';
+import { truncateHash } from '../utils/helpers';
 
 interface InitialState {
   activeComparison: string;
@@ -18,6 +19,15 @@ type Results = {
   revisionHeader: RevisionsHeader;
 };
 
+interface GroupedResults {
+  [key: string]: ResultObject[];
+}
+
+interface ResultObject {
+  header_name: string;
+  new_rev: string;
+}
+
 const comparison = createSlice({
   name: 'comparison',
   initialState,
@@ -32,6 +42,56 @@ const comparison = createSlice({
     },
   },
 });
+
+export const formatDownloadData = (
+  data: Record<string, ResultObject[]>,
+): object[] => {
+  const groupNames = Object.keys(data);
+  const transformedGroups: object[] = [];
+
+  groupNames.forEach((groupName) => {
+    const groupedResults = data[groupName].reduce(
+      (grouped: GroupedResults, result: ResultObject) => {
+        if (!grouped[result.header_name]) {
+          grouped[result.header_name] = [];
+        }
+        grouped[result.header_name].push(result);
+        return grouped;
+      },
+      {},
+    );
+
+    const transformedGroupEntries = Object.keys(groupedResults).map(
+      (header_name) => ({
+        [`${header_name} ${truncateHash(
+          groupedResults[header_name][0].new_rev,
+        )}`]: groupedResults[header_name],
+      }),
+    );
+
+    transformedGroups.push(...transformedGroupEntries);
+  });
+  return transformedGroups;
+};
+
+export const selectStringifiedJsonResults = createSelector(
+  (state: RootState) => state.comparison.activeComparison,
+  (state: RootState) => state.compareResults.data,
+  (activeComparison, data) => {
+    if (
+      activeComparison ===
+      Strings.components.comparisonRevisionDropdown.allRevisions
+    ) {
+      return JSON.stringify(formatDownloadData(data));
+    } else {
+      return JSON.stringify(
+        formatDownloadData({
+          [activeComparison]: data[activeComparison],
+        }),
+      );
+    }
+  },
+);
 
 function processResults(results: CompareResultsItem[]) {
   const processedResults: Map<string, CompareResultsItem[]> = new Map<
